Refactor. 폴더구조 변경 - js폴더로 모든 js파일 모음 - pages 폴더 밑에 하위폴더 제거 , html파일을 바로 배치
