+++ src/pages/detail.html
@@ -17,6 +17,6 @@
       <div class="detail__movie-info-wrapper"></div>
     </main>
     <footer></footer>
-    <script src="script.js" type="module"></script>
+    <script src="/src/js/detail.js" type="module"></script>
   </body>
 </html>
