+++ src/pages/header.html
@@ -21,4 +21,4 @@
     />
   </div>
 </header>
-<script src="/src/headerScript.js" type="module"></script>
+<script src="/src/js/header.js" type="module"></script>
