--- src/pages/search/_styles.scss
@@ -1,1 +1,0 @@
-
