+++ src/pages/searchResult.html
@@ -36,6 +36,6 @@
       <h1 class="recommend__title">RECOMMEND</h1>
       <div class="recommend__content"></div>
     </section>
-    <script type="module" src="script.js"></script>
+    <script type="module" src="/src/js/searchResult.js"></script>
   </body>
 </html>
